+++ src/plugin.js
@@ -17,6 +17,10 @@ const defaults = {
 
 const Component = videojs.getComponent('Component');
 
+// These are for cross-compatibility between Video.js 5 and 6.
+const dom = videojs.dom || videojs;
+const registerPlugin = videojs.registerPlugin || videojs.plugin;
+
 /**
  * Whether the value is a `Number`.
  *
@@ -93,7 +97,7 @@ class Overlay extends Component {
     let content = options.content;
 
     let background = options.showBackground ? 'vjs-overlay-background' : 'vjs-overlay-no-background';
-    let el = videojs.createEl('div', {
+    let el = dom.createEl('div', {
       className: `
         vjs-overlay
         vjs-overlay-${options.align}
@@ -108,7 +112,7 @@ class Overlay extends Component {
     } else if (content instanceof window.DocumentFragment) {
       el.appendChild(content);
     } else {
-      videojs.appendContent(el, content);
+      dom.appendContent(el, content);
     }
 
     return el;
@@ -342,6 +346,6 @@ const plugin = function(options) {
 
 plugin.VERSION = '__VERSION__';
 
-videojs.plugin('overlay', plugin);
+registerPlugin('overlay', plugin);
 
 export default plugin;